Add Run Test Scenario example
It is possible to work with Scesim files
inside VSCode.
Test Scenario Simulation works for DMN and awalible to use
if user includes Kogito ScenarioJunitActivator and
adds related scesim file to the project + add kogito scenario simulation dependency.

I made this example to look exactly as it is implemented in Business central to make it familiar for users.